+++ CHANGELOG.md
@@ -12,6 +12,7 @@ FEATURES:
 * `vcd_vapp_network` supports isolated network and vApp network connected to Org VDC networks [GH-455]
 * **New Data Source:** `vcd_vapp_org_network` vApp org network [GH-455]
 * **New Data Source:** `vcd_vapp_network` vApp network [GH-455]
+* `resource/vcd_vapp_vm` and `datasource/vcd_vapp_vm` `customization` block supports all available features [GH-462]
 
 IMPROVEMENTS:
 
@@ -35,6 +36,12 @@ BUG FIXES:
 * `resource/vcd_vapp_vm` `network` block changes caused MAC address changes in existing NICs
   [GH-436,GH-407]
 * Fix a potential data race in client connection caching when VCD_CACHE is enabled [GH-453]
+* `resource/vcd_vapp_vm` when customization.0.force=false crashes with interface {} is nil [GH-462]
+* `resource/vcd_vapp_vm` `customization.0.force=true` could have skipped "Forced customization" on each apply [GH-462]
+
+DEPRECATIONS:
+
+* `resource/vcd_vapp_vm` field `initscript` is now deprecated in favor of `customization.0.initscript` [GH-462]
 
 NOTES:
 
@@ -42,6 +49,7 @@ NOTES:
 * Bump terraform-plugin-sdk to v1.5.0 [GH-442]
 * `make seqtestacc` and `make test-binary` use `-race` flags for `go test` to check if there are no data races.
  Additionally GNUMakefile supports `make installrace` and `make buildrace` to build binary with race detection enabled. [GH-453]
+* Added `make test-upgrade-prepare` directive [GH-462]
 
 ## 2.6.0 (December 13, 2019)